rework VCO tuning
The current VCO tuning algorithm experiences periodic failures. We have isolated these to
"stuck" comparator readings at the low and high end of the CSW range, while a valid
solution exists inside the range. We change the algorithm to exhaustively search the space
and debounce the reading to ensure we can always find a valid CSW value and avoid failure.
Root cause of the "stuck" comparator is TBD.